// types.ts
export interface Npc {
  name: string;
  gender: string;
  race: string;
  age: string;
  intelligence: string;
  hairStyle: string;
  hairColor: string;
  facialHair: string;
  height: string;
  weight: string;
  eyeShape: string;
  eyeColor: string;
  complexion: string;
  descriptor: string;
  profession: string;
  demeanor: string;
  wantsOrNeed: string;
  secretOrObstacle: string;
  alsoCarrying: string[];
  gold: number;
  silver: number;
  copper: number;
  fullDescription: string;
}